refactor: move Mint and Melt quote to cdk common
