+++ crates/cdk-common/src/lib.rs
@@ -10,6 +10,8 @@ pub mod common;
 pub mod database;
 pub mod error;
 #[cfg(feature = "mint")]
+pub mod mint;
+#[cfg(feature = "mint")]
 pub mod payment;
 pub mod pub_sub;
 pub mod subscription;
@@ -19,8 +21,6 @@ pub mod ws;
 pub use bitcoin;
 pub use cashu::amount::{self, Amount};
 pub use cashu::lightning_invoice::{self, Bolt11Invoice};
-#[cfg(feature = "mint")]
-pub use cashu::mint;
 pub use cashu::nuts::{self, *};
 #[cfg(feature = "wallet")]
 pub use cashu::wallet;
